+++ packages/synthesis/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @claudexor/synthesis
 
+## 3.10.4
+
+### Patch Changes
+
+- @claudexor/arbitration@3.10.4
+
 ## 3.10.3
 
 ### Patch Changes